pro/django_homework/project_hw/requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.3
celery==5.5.3
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.12.0
cryptography==46.0.3
Django==5.2.7
django-debug-toolbar==6.1.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==38.2.0
iniconfig==2.3.0
kombu==5.5.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
